cli_monitor/common/config.py
import os
from dotenv import load_dotenv

load_dotenv()

API_KEY = os.getenv("BINANCE_API_KEY")
API_SECRET = os.getenv("BINANCE_API_SECRET")

if not API_KEY or not API_SECRET:
    raise ValueError("Binance API keys not found in .env file. Please set BINANCE_API_KEY and BINANCE_API_SECRET.")
